// web/src/api/models/RenderNftVar.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RenderNftVar
 */
export interface RenderNftVar {
    /**
     * 
     * @type {string}
     * @memberof RenderNftVar
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof RenderNftVar
     */
    value: string;
}

/**
 * Check if a given object implements the RenderNftVar interface.
 */
export function instanceOfRenderNftVar(value: object): value is RenderNftVar {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('value' in value) || value['value'] === undefined) return false;
    return true;
}

export function RenderNftVarFromJSON(json: any): RenderNftVar {
    return RenderNftVarFromJSONTyped(json, false);
}

export function RenderNftVarFromJSONTyped(json: any, ignoreDiscriminator: boolean): RenderNftVar {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'value': json['value'],
    };
}

export function RenderNftVarToJSON(json: any): RenderNftVar {
    return RenderNftVarToJSONTyped(json, false);
}

export function RenderNftVarToJSONTyped(value?: RenderNftVar | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'value': value['value'],
    };
}